requirements.txt
python-multipart
sqlalchemy
jwt
fastapi
python-dotenv
uvicorn
mypy
google-cloud-storage
bcrypt
psycopg2-binary
pydantic